refactor: updated strategy pattern UML diagram
