+++ src/backend/main/gamehistory.js
@@ -1,4 +1,3 @@
-
 const games = [];
 
 function save(game) {
